+++ tunnel/transport/server.go
@@ -28,7 +28,7 @@ type Server struct {
 
 func (s *Server) Close() error {
 	s.cancel()
-	if s.cmd != nil {
+	if s.cmd != nil && s.cmd.Process != nil {
 		s.cmd.Process.Kill()
 	}
 	return s.tcpListener.Close()
